requirements.txt
chardet==3.0.4
click==7.1.2
dominate==2.5.2
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.4.1
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==1.0.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
requests==2.24.0
SQLAlchemy==1.3.19
urllib3==1.25.10
visitor==0.1.3
Werkzeug==2.0.0
WTForms==2.3.3
Pillow==9.1.0
numpy==1.22.3
scipy==1.8.0
yagmail==0.15.277